# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Courtside TW API
  version: 1.0.0
  description: Contract-first REST API for public reading, reader library, editorial publication and optional provenance/SIWE boundaries.
servers:
  - url: https://api.courtside.tw
    description: Production API origin placeholder; deployment is a later task.
...
security:
  - bffSession: []
  - oidcBearer: []
paths:
...
  /api/v1/editor/articles:
    post:
      tags:
        - Editorial
      operationId: createEditorArticle
      summary: Create an article draft
      description: Creates article identity and an initial working revision.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
              examples:
                valid:
                  value:
                    articleId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    revisionNumber: 1
                    version: 1
                    title: Article 1
                    slug: article-1
                    state: DRAFT
                    readiness:
                      ready: true
                      blockingCodes: []
                      blockers: []
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleDraftInput"
            examples:
              valid:
                value:
                  title: Article 1
                  slug: article-1
      x-idempotent: true
      x-required-roles:
        - EDITOR
    get:
      tags:
        - Editorial
      operationId: listEditorArticles
      summary: List article drafts
      description: Lists article identities and their current editor-visible revisions.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraftPage"
              examples:
                valid:
                  value:
                    items: []
                    page:
                      nextCursor: null
                      limit: 20
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - EDITOR
    patch:
      tags:
        - Editorial
      operationId: patchEditorArticle
      summary: Patch an article draft
      description: Updates an article identity or working revision with an If-Match lock.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
              examples:
                valid:
                  value:
                    articleId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    revisionNumber: 1
                    version: 2
                    title: Article 1 revised
                    slug: article-1
                    state: DRAFT
                    readiness:
                      ready: true
                      blockingCodes: []
                      blockers: []
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticlePatch"
            examples:
              valid:
                value:
                  articleId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                  changes:
                    title: Article 1 revised
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
  /api/v1/editor/articles/{id}:submit:
    post:
      tags:
        - Editorial
      operationId: submitArticleForReview
      summary: Submit an article revision
      description: Validates content and rights before moving the article into review.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResult"
              examples:
                valid:
                  value:
                    operationId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    status: ACCEPTED
                    version: 1
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitRequest"
            examples:
              valid:
                value:
                  revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
      x-idempotent: true
      x-required-roles:
        - EDITOR
  /api/v1/editor/articles/{id}:revise:
    post:
      tags:
        - Editorial
      operationId: createArticleRevision
      summary: Create an article revision
      description: Creates a new draft revision without mutating the published snapshot.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Revision created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
              examples:
                valid:
                  value:
                    articleId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    revisionNumber: 1
                    version: 2
                    title: Article correction
                    slug: article-1
                    state: DRAFT
                    readiness:
                      ready: true
                      blockingCodes: []
                      blockers: []
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevisionInput"
            examples:
              valid:
                value:
                  title: Article correction
                  dek: Corrected deck
                  content:
                    schemaVersion: 1
                    documentId: 0190f7b0-7c4b-7e3a-8f12-123456789abe
                    blocks:
                      - id: 00000000-0000-4000-8000-000000000102
                        type: paragraph
                        version: 1
                        payload:
                          content:
                            - kind: text
                              text: Article correction body
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - EDITOR
  /api/v1/publisher/articles/{id}:approve:
    post:
      tags:
        - Publisher
      operationId: approvePublisherArticle
      summary: Approve a frozen revision
      description: Approves an article revision after editorial review.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResult"
              examples:
                valid:
                  value:
                    operationId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    status: APPROVED
                    version: 1
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
        "429":
          $ref: "#/components/responses/Problem429"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - PUBLISHER
  /api/v1/publisher/articles:
    get:
      tags:
        - Publisher
      operationId: listPublisherArticles
      summary: List articles for publisher review
      description: Lists article revisions and readiness evidence visible to a publisher.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraftPage"
              examples:
                valid:
                  value:
                    items: []
                    page:
                      nextCursor: null
                      limit: 20
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - PUBLISHER
  /api/v1/publisher/articles/{id}:
    get:
      tags:
        - Publisher
      operationId: getPublisherArticle
      summary: Read an article for publisher review
      description: Returns the current article revision and its readiness evidence.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Successful response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleDraft"
              examples:
                valid:
                  value:
                    articleId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    revisionNumber: 1
                    version: 1
                    title: Article 1
                    slug: article-1
                    state: DRAFT
                    readiness:
                      ready: true
                      blockingCodes: []
                      blockers: []
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "429":
          $ref: "#/components/responses/Problem429"
      x-required-roles:
        - PUBLISHER
  /api/v1/publisher/articles/{id}:schedule:
    post:
      tags:
        - Publisher
      operationId: schedulePublisherArticle
      summary: Schedule an article
      description: Schedules article publication using an explicit IANA timezone.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleRequest"
            examples:
              valid:
                value:
                  publishAt: 2026-08-11T09:00:00
                  timezone: Asia/Taipei
      responses:
        "202":
          description: Schedule accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResult"
              examples:
                valid:
                  value:
                    operationId: 0190f7b0-7c4b-7e3a-8f12-123456789abc
                    status: SCHEDULED
                    version: 4
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
                    scheduledAt: 2026-08-11T01:00:00Z
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - PUBLISHER
  /api/v1/publisher/articles/{id}:publish:
    post:
      tags:
        - Publisher
      operationId: publishPublisherArticle
      summary: Publish an article now
      description: Publishes the approved article revision and records an immutable snapshot.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Publication accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResult"
              examples:
                valid:
                  value:
                    operationId: 0190f7b0-7c4b-7e3a-8f12-123456789abe
                    status: PUBLISHED
                    version: 5
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - PUBLISHER
  /api/v1/publisher/articles/{id}:request-changes:
    post:
      tags:
        - Publisher
      operationId: requestChangesPublisherArticle
      summary: Request article changes
      description: Returns a frozen article revision to draft with an auditable reason.
      security:
        - bffSession: []
        - oidcBearer: []
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/IfMatch"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawRequest"
            examples:
              valid:
                value:
                  reason: editorial changes requested
      responses:
        "202":
          description: Change request accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowResult"
              examples:
                valid:
                  value:
                    operationId: 0190f7b0-7c4b-7e3a-8f12-123456789abf
                    status: CHANGES_REQUESTED
                    version: 4
                    revisionId: 0190f7b0-7c4b-7e3a-8f12-123456789abd
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
        "400":
          $ref: "#/components/responses/Problem400"
        "401":
          $ref: "#/components/responses/Problem401"
        "403":
          $ref: "#/components/responses/Problem403"
        "404":
          $ref: "#/components/responses/Problem404"
        "409":
          $ref: "#/components/responses/Problem409"
        "422":
          $ref: "#/components/responses/Problem422"
      x-idempotent: true
      x-optimistic-lock: If-Match
      x-required-roles:
        - PUBLISHER
...
components:
  securitySchemes:
    oidcBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC issuer/audience validated JWT for Spring Security.
    bffSession:
      type: apiKey
      in: cookie
      name: __Host-courtside_session
      description: Secure, HttpOnly, SameSite BFF session cookie.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        type: string
        minLength: 1
        maxLength: 256
...
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Stable retry key. Replays return the original operation result.
      schema:
        type: string
        format: uuid
        minLength: 36
        maxLength: 200
  headers:
    XRequestId:
      description: Trace/request identifier. Never contains tokens or article content.
      schema:
        type: string
        minLength: 1
        maxLength: 100
...
  schemas:
...
    ContentDocument:
      $ref: ./content-document.schema.json
...
    ArticleDraft:
      type: object
      additionalProperties: false
      required:
        - articleId
        - revisionId
        - revisionNumber
        - version
        - title
        - slug
        - state
        - readiness
      properties:
        articleId:
          $ref: "#/components/schemas/Uuid"
        revisionId:
          $ref: "#/components/schemas/Uuid"
        revisionNumber:
          type: integer
          minimum: 1
        version:
          type: integer
          minimum: 1
        title:
          type: string
        slug:
          type: string
        dek:
          type: string
        content:
          $ref: "#/components/schemas/ContentDocument"
        state:
          type: string
          enum:
            - DRAFT
            - IN_REVIEW
            - APPROVED
            - SCHEDULED
            - PUBLISHED
            - WITHDRAWN
            - ARCHIVED
        scheduledAt:
          type: string
          format: date-time
        readiness:
          $ref: "#/components/schemas/PublicationReadiness"
...
  responses:
    Problem400:
      description: Invalid request (400).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          examples:
            stable:
              summary: INVALID_REQUEST example
              value:
                type: https://courtside.tw/problems/invalid_request
                title: Invalid request
                status: 400
                detail: The request is invalid.
                instance: /api/v1/example
                requestId: req_t009_t011_example
                code: INVALID_REQUEST
      x-stable-error-code: INVALID_REQUEST
...